pyproject.toml
[project]
name = "agentx-purple-agent"
version = "0.1.0"
description = "A scaffolded AgentX Kubernetes purple-team A2A agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20,<1.0.0",
    "litellm>=1.83.13",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
